rename userspace files and fix includes
