+++ keyboards/bastardkb/charybdis/3x5/keymaps/mute-civilian/keymap.c
@@ -1,4 +1,4 @@
-#include "nathancooke7.h"
+#include "mutecivilian.h"
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
   [_BASE] = LAYOUT_charybdis_3x5(
