+++ users/mute-civilian/features/secrets.c
@@ -1,4 +1,4 @@
-#include "nathancooke7.h"  // replace with your keymap's "h" file, or whatever file stores the keycodes
+#include "mutecivilian.h"  // replace with your keymap's "h" file, or whatever file stores the keycodes
 
 #if (__has_include("features/secrets.h") && !defined(NO_SECRETS))
 #include "features/secrets.h"
