+++ users/mute-civilian/mutecivilian.c
@@ -1,4 +1,4 @@
-#include "nathancooke7.h"
+#include "mutecivilian.h"
 
 #ifdef TAPPING_TERM_PER_KEY
 uint16_t get_tapping_term(uint16_t keycode, keyrecord_t *record) {
